fix: grind GameObject IDs, AI objects, Map, PauseWindow, Visos
Implement STUB-to-FUNCTION bodies across AI (GameObject id bitmap,
bbox, destinations, object types), Map, PauseWindow, VsTrig, BitmapRes,
and assorted Visos network/resource near-miss fixes for reccmp parity.

+++ src/Visos/Animation/TimedAnim.cpp
@@ -34,7 +34,7 @@ unsigned int TimedAnim::GetFrameNo()
 		time = timeGetTime();
 	}
 	time -= m_frameState;
-	unsigned int frame = ((time % m_animTime) * m_frames) / m_animTime;
+	unsigned long frame = ((time % m_animTime) * m_frames) / m_animTime;
 	if (m_direction != 1) {
 		frame = m_frames - frame - 1;
 	}
